Explicited a cast from double to uint32_t when updating the sampleRate
Note that if sampleRate*m_minSmoothingDuration is not an integer, then the static_cast will truncate the number of samples to use for parameter smoothing. To be more accurate, we would therefore need an extra sample to properly cover the whole smoothing duration, but this was not implemented since this truncation would rarely happen in practice (if m_minSmoothingDuration is equal to an integer when expressed in milliseconds and sampleRate is equal to an integer when expressed in kHz, which is usually the case, then the multiplication will result in an integer).

+++ source/core/instrument/Instrument.cpp
@@ -56,7 +56,9 @@ namespace ANGLECORE
 
     void BaseInstrument::ParameterRenderer::setSampleRate(double sampleRate)
     {
-        m_minSmoothingSamples.store(m_minSmoothingDuration * sampleRate);
+        m_minSmoothingSamples.store(static_cast<uint32_t>(m_minSmoothingDuration * sampleRate));
+        /* Note that we should actually add an extra sample when the result of the
+        * multiplication is not an integer (which should be rare) */
     }
 
     void BaseInstrument::ParameterRenderer::setMaxSamplesPerBlock(uint32_t maxSamplesPerBlock)
